Refactor SystematicReviewAgent: extract exceptions and resume logic
Major refactoring to reduce agent.py from 3,087 to 1,829 lines (41% reduction):

- Create exceptions.py: All exception classes (66 lines)
  - SystematicReviewError and subclasses
  - New CheckpointError and ResumeError exceptions

- Create resume_mixin.py: Checkpoint resume functionality (1,336 lines)
  - CheckpointResumeMixin class with all resume methods
  - run_review_from_checkpoint()
  - _restore_state_from_checkpoint()
  - All _continue_from_* continuation methods

- Update agent.py:
  - Import exceptions from exceptions.py
  - Inherit from CheckpointResumeMixin
  - Remove inline exception definitions
  - Remove resume methods (now in mixin)

- Update __init__.py:
  - Export CheckpointResumeMixin
  - Export new CheckpointError and ResumeError
  - Import exceptions from new module

All tests pass. Module structure preserved with clean separation of concerns.
